Add Lottie assets and markers support, update timeline types to Float, and document serialization types
- Add polymorphic Asset models (PrecompAsset, ImageAsset, AudioAsset, UnknownAsset) supporting embedded Base64 data URIs and subcompositions.
- Add Marker data model supporting temporal cue points and comments.
- Update Animation header timeline fields (fr, ip, op) from Int to Float aligning with Lottie 1.0.1 specification and fractional frame rates.
- Add comprehensive serialization documentation in remotecompose/lottie/docs/lottie_specification_types.md.
- Add unit tests in AssetTest and ParsingTest.

TAG=agy
CONV=6430017c-8336-4849-a261-1ab75c80d930

+++ remotecompose/lottie/src/debug/java/com/google/android/horologist/remotecompose/lottie/LottiePreview.kt
@@ -130,7 +130,7 @@ internal fun LottieAnimatedPreview(
   val duration =
     durationMillis
       ?: if (animation.frameRate > 0) {
-        (((animation.endFrame - animation.startFrame) / animation.frameRate.toFloat()) * 1000)
+        (((animation.endFrame - animation.startFrame) / animation.frameRate) * 1000)
           .toInt()
           .coerceAtLeast(100)
       } else {
